+++ src/main/java/objects/interfaces/Damaging.java
@@ -2,6 +2,6 @@ package objects.interfaces;
 
 public interface Damaging {
 
-    void setDamage(double damage);
-    void getDamage(double damage);
+    void setDamage(float damage);
+    void getDamage(float damage);
 }
